relocated mixin to proper package + other general fixes
